feat(errors): implementar formato RFC 7807 Problem Details
Modificar error_response() para retornar estructura estándar:
{type, title, status, detail, code, field}

- Agregar parámetros code (machine-readable) y field (campo afectado)
- Actualizar todos los routers para pasar code/field en cada error
- BusinessRuleError para errores de negocio en services
- Mantener success responses con ApiResponse sin cambios

+++ backend/app/modules/pagos/schema.py
@@ -1,11 +1,12 @@
 from typing import Optional
+from decimal import Decimal
 from datetime import datetime
 from sqlmodel import SQLModel
 
 
 class PagoCreate(SQLModel):
     pedido_id: int
-    transaction_amount: float
+    transaction_amount: Decimal
 
 
 class PagoRead(SQLModel):
@@ -14,7 +15,7 @@ class PagoRead(SQLModel):
     mp_payment_id: Optional[int] = None
     mp_status: str
     mp_status_detail: Optional[str] = None
-    transaction_amount: float
+    transaction_amount: Decimal
     external_reference: str
     idempotency_key: str
     payment_method_id: Optional[str] = None
